requirements.txt
asgiref==3.2.10
certifi==2020.11.8
chardet==3.0.4
Django==3.1.2
idna==2.10
Pillow==8.0.1
pytz==2020.1
requests==2.25.0
sqlparse==0.4.1
stripe==2.55.1
urllib3==1.26.2
